Added semi collision, and thunderkit pipeline shit
The pipeline now deletes the current bundle, if it's there, which means it lets you rebuild the bundle real easily. Gonna add copying to game folder later but lazy

Oh and also added bad collision to the seal. No clue if this is good way to do it or not it was just easy