Shite
Moved recipes to load from Recipes.json file, easier to edit as a non programmer (or even as a programmer) and more readable

Added pda groups and categories for the seal & upgrades, along with a patch to make them appear in place directly below the cyclops in the menu, rather than all the way at the bottom

WIP bug fix for base pieces. Just got back to project from being sick for a week + can't remember how much is or isn't functional. Will follow up with a commit to fix shit if it's *too* broken